Add files via upload
DNS is a service that is used on every system with general purpose use of the internet. It is therefore a convenient
target for misuse. The misuse under consideration here is tunneling. Malicious users extract confidential data from
compromised computers by encoding the data to be tunneled within a DNS query name while avoiding firewalls
and other network security monitoring mechanisms. We developed machine learning (ML) classifiers for the
detection of DNS Tunneling attacks.
--------------------------------------------------------------------------------------------------------------